Authentication implemented, and some fixes in crud

+++ controllers/feed.js
@@ -1,5 +1,6 @@
 const { validationResult } = require("express-validator/check");
 const Product = require("../models/product");
+const User = require("../models/user");
 const fs = require('fs');
 const path = require('path');
 
@@ -36,6 +37,7 @@ exports.createPost = (req, res, next) => {
   const code = req.body.code;
   // const imageUrl = req.body.imageUrl;
   const sizes = req.body.sizes;
+  let creator;
 
   const product = new Product({
     title: title,
@@ -45,14 +47,25 @@ exports.createPost = (req, res, next) => {
       imageUrl: imageUrl,
       sizes: sizes,
     },
+    creator: req.userId
   });
   product
     .save()
     .then((result) => {
-      console.log(result);
+      return User.findById(req.userId);
+    })
+    .then(user => {
+      creator = user;
+      user.products.push(product)
+      return user.save();
+    })
+    .then(result  => {
       res.status(201).json({
         message: "Product created successfully!",
-        product: result,
+        product: product,
+        creator: {
+          _id: creator._id,
+        }
       });
     })
     .catch((err) => {
@@ -60,7 +73,7 @@ exports.createPost = (req, res, next) => {
         err.statusCode = 500;
       }
       next(err);
-    }); 
+    });
 };
 
 exports.getPost = (req, res, rext) => {
@@ -92,8 +105,10 @@ exports.updatePost = (req, res, next) => {
     throw error;
   }
   const title = req.body.title;
-  const content = req.body.content;
-  let imageUrl = req.body.imageUrl; // imageUrl should be as in frontend
+  const price = req.body.price;
+  const code = req.body.code;
+  const imageUrl = req.body.imageUrl; // imageUrl should be as in frontend
+  const sizes = req.body.sizes;
 
   if(req.file) {
     imageUrl = req.file.path;
@@ -110,17 +125,25 @@ exports.updatePost = (req, res, next) => {
       error.statusCode = 404;
       throw error;
     }
+
+    if(product.creator.toString() !== req.userId) {
+      const error = new Error('Not authorized!');
+      error.statusCode = 403;
+      throw error;
+    }
     if(imageUrl !== product.imageUrl) {
         clearImage(product.imageUrl);
     }
 
     product.title = title;
-    product.imageUrl = imageUrl;
-    product.content = content;
-    return product.save();
+      product.price = price;
+      product.code = code;
+      product.sizes = sizes;
+      product.imageUrl = imageUrl;
+      return product.save();
   })
-  .then(result => {
-    res.status(200).json({message: 'Product updated!', product: result})
+  .then((result) => {
+    res.status(200).json({ message: "Product updated!", product: result });
   })
   .catch((err) => {
     if (!err.statusCode) {
@@ -133,26 +156,37 @@ exports.updatePost = (req, res, next) => {
 exports.deletePost = (req, res, next) => {
   const productId = req.params.productId;
   Product.findById(productId)
-  .then(product => {
-    if(!product) {
-      const error = new Error('Could not find product.');
-      error.statusCode = 404;
-      throw error;
-    }
-    // Check logged user
-    // clearImage(product.imageUrl);
-    return Product.findByIdAndRemove(productId);
-  })
-  .then(result => {
-    console.log(result);
-    res.status(200).json({message: 'Deleted product.'});
-  })
-  .catch((err) => {
-    if (!err.statusCode) {
-      err.statusCode = 500;
-    }
-    next(err);
-  });
+    .then((product) => {
+      if (!product) {
+        const error = new Error("Could not find product.");
+        error.statusCode = 404;
+        throw error;
+      }
+      // Check logged user
+      if(product.creator.toString() !== req.userId) {
+        const error = new Error('Not authorized!');
+        error.statusCode = 403;
+        throw error;
+      }
+      clearImage(product.imageUrl);
+      return Product.findByIdAndRemove(productId);
+    })
+    .then((result) => {
+      return User.findById(req.userId);
+    })
+    .then(user => {
+      user.products.pull(productId)
+      return user.save();
+    })
+    .then(result => {
+      res.status(200).json({ message: "Deleted product." });
+    })
+    .catch((err) => {
+      if (!err.statusCode) {
+        err.statusCode = 500;
+      }
+      next(err);
+    });
 };
 
 const clearImage = filePath => {
